refactor: move util function to component + minor changes to label and description
